requirements.txt
# BTC Sniper Bot - Dependencies
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
requests>=2.31.0
